scripts/pyproject.toml
[project]
name = "kanji-verse-scripts"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "sentence-transformers>=2.7.0",
    "umap-learn>=0.5.6",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
]
